fix(CartManager): 가격은 double임...

+++ app/src/main/java/kiosk/managers/CartManager.java
@@ -13,7 +13,7 @@ public class CartManager {
             // (oldValue, newValue) -> oldValue + newValue랑 같은 의미
     }
 
-    public int getTotalPrice() {
+    public double getTotalPrice() {
         // entrySet 또한 stream이 가능함. 처음 알았음.
         return cartItems.entrySet().stream()
                 .mapToInt(entry -> (int) (entry.getKey().price() * entry.getValue()))
